图片模块：用户推荐-获取需要推荐的标签和分类分数

+++ Picture/picture-picture/src/main/java/com/lz/picture/model/domain/PictureDownloadLogInfo.java
@@ -157,7 +157,7 @@ public class PictureDownloadLogInfo implements Serializable {
      * 分数
      */
     @Excel(name = "分数")
-    private BigDecimal score;
+    private Double score;
 
     /**
      * IP地址
